+++ examples/web/src/main.ts
@@ -5,6 +5,7 @@ import App from './App.vue'
 import ApiClientPage from './pages/ApiClientPage.vue'
 import ApiReferencePage from './pages/ApiReferencePage.vue'
 import ClassicApiReferencePage from './pages/ClassicApiReferencePage.vue'
+import EditableApiReferencePage from './pages/EditableApiReferencePage.vue'
 import StandaloneApiReferencePage from './pages/StandaloneApiReferencePage.vue'
 import StartPage from './pages/StartPage.vue'
 import SwaggerEditorPage from './pages/SwaggerEditorPage.vue'
@@ -27,6 +28,11 @@ const routes = [
     path: '/classic-api-reference',
     name: 'classic-api-reference',
     component: ClassicApiReferencePage,
+  },
+  {
+    path: '/editable-api-reference',
+    name: 'editable-api-reference',
+    component: EditableApiReferencePage,
   },
   {
     path: '/swagger-editor',