+++ examples/web/src/pages/EditableApiReferencePage.vue
@@ -6,13 +6,11 @@ import {
 import { reactive } from 'vue'
 
 import SlotPlaceholder from '../components/SlotPlaceholder.vue'
-import content from '../fixtures/petstorev3.json'
 
 const configuration = reactive<ReferenceConfiguration>({
   theme: 'default',
   proxy: 'http://localhost:5051',
-  isEditable: false,
-  spec: { content },
+  isEditable: true,
 })
 </script>
 <template>